Finalized the app conversion
