+++ apps/admin/forms/admin.php
@@ -6,8 +6,8 @@
     $pass = $_POST['adPass'];
     $logout = $_POST['adLogout'];
 
-    // Because this runs from a subdir /root/templates/forms
-    $root = str_replace('templates/forms', '', dirname(__FILE__));
+    // Because this runs from a subdir
+    $root = str_replace('apps/admin/forms', '', dirname(__FILE__));
 
     // Only process if the form was actually submitted
     if (isset($user) && isset($pass)) {
